refactor(096-W3): 黑板 flush 函数族搬离 completion.rs → services/blackboard_flush.rs（-370 行）
- Move Function 整族搬迁（函数体零改动）：blackboard_flush_listener / handle_flush_msg /
  spawn_flush_worker / broadcast_ticker_status / build_blackboard_status / get_workspace_debounce
- 消除跨层倒挂：listener 原为 pub 被 handlers 召唤，迁后落黑板服务域并收紧为 pub(crate)
- completion.rs 1081 → 711 行；record 入队动作（push_pending_record）属执行域留在 finalize
- clippy 零告警、全量 1676 测试过
